+++ hutool-core/src/main/java/com/xiaoleilu/hutool/getter/OptNullBasicTypeFromObjectGetter.java
@@ -7,7 +7,7 @@ import com.xiaoleilu.hutool.convert.Convert;
 
 /**
  * 基本类型的getter接口抽象实现，所有类型的值获取都是通过将getObj获得的值转换而来<br>
- * 用户只需实现getStr方法即可，其他类型将会从String结果中转换
+ * 用户只需实现getObj方法即可，其他类型将会从Object结果中转换
  * 在不提供默认值的情况下， 如果值不存在或获取错误，返回null<br>
  * @author Looly
  */